neurax_complet/neurax_complet/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "h5py>=3.12.1",
    "matplotlib>=3.10.0",
    "numpy>=2.2.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "scipy>=1.15.2",
    "streamlit>=1.42.0",
    "trafilatura>=2.0.0",
    "twilio>=9.4.5",
]
